+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 * Makes I18NextOption conform to Map<String, Object>
 * Reduces API surface by merging most of the optional properties into I18NextOptions itself
 * Moves pattern builders from options to the classes themselves
+* Keeps property variables in I18NextOptions while keeping Map extension.
 
 ## [0.0.1+3]